+++ spec/Gnugat/Medio/Service/CodeDetectorSpec.php
@@ -53,6 +53,20 @@ class CodeDetectorSpec extends ObjectBehavior
         $this->isUseNeeded($text, self::NAME_SPACE)->shouldBe(true);
     }
 
+    function it_detects_presence_of_method(Editor $editor, Text $text)
+    {
+        $editor->hasBelow($text, self::METHOD_PATTERN, 0)->willReturn(true);
+
+        $this->hasMethod($text, self::METHOD_NAME)->shouldBe(true);
+    }
+
+    function it_detects_absence_of_method(Editor $editor, Text $text)
+    {
+        $editor->hasBelow($text, self::METHOD_NAME, 0)->willReturn(false);
+
+        $this->hasMethod($text, self::METHOD_NAME)->shouldBe(false);
+    }
+
     function it_detects_arguments_of_multiline_method(
         CodeNavigator $codeNavigator,
         Editor $editor,