+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.3.0
+
+* added detection of spec method
+
 ## 0.2.0
 
 * added insertion of missing constructor in dependency injection command